added CreationResponseDTO refactored createUser()
