+++ src/main/java/fr/oukilson/backend/configuration/UserConfiguration.java
@@ -5,6 +5,7 @@ import fr.oukilson.backend.service.UserService;
 import io.github.cdimascio.dotenv.Dotenv;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
@@ -13,10 +14,15 @@ import java.util.regex.Pattern;
 @Configuration
 public class UserConfiguration {
 
-    /*
-    create a dotenv object to retrieve the environmental variables
-     */
-    Dotenv dotenv = Dotenv.load();
+
+    @Value("${environment.emailRegex}")
+    private String emailRegex;
+
+    @Value("${environment.nicknameRegex}")
+    private String nicknameRegex;
+
+    @Value("${environment.nameRegex}")
+    private String nameRegex;
 
     /*
     create a service bean to use inside the user side of the project
@@ -31,16 +37,16 @@ public class UserConfiguration {
 
     @Bean
     public Pattern emailPattern(){
-        return Pattern.compile(dotenv.get("EMAIL_REGEX"));
+        return Pattern.compile(emailRegex);
     }
 
     @Bean
     public Pattern nicknamePattern(){
-        return Pattern.compile(dotenv.get("NICKNAME_REGEX"));
+        return Pattern.compile(nicknameRegex);
     }
 
     @Bean
     public Pattern namePattern(){
-        return Pattern.compile(dotenv.get("NAME_REGEX"));
+        return Pattern.compile(nameRegex);
     }
 }
