+++ src/main/java/fr/oukilson/backend/controller/UserController.java
@@ -1,6 +1,7 @@
 package fr.oukilson.backend.controller;
 
 import fr.oukilson.backend.dto.UserCreationDTO;
+import fr.oukilson.backend.dto.CreationResponseDTO;
 import fr.oukilson.backend.service.UserService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -26,7 +27,7 @@ public class UserController {
      * @return a response entity validating the created user
      */
     @PostMapping()
-    public ResponseEntity<UserCreationDTO> createUser(@RequestBody UserCreationDTO userCreationDTO){
+    public ResponseEntity<CreationResponseDTO> createUser(@RequestBody UserCreationDTO userCreationDTO){
         return ResponseEntity.status(HttpStatus.CREATED).body(this.userService.createUser(userCreationDTO));
     }
 }
