+++ src/main/java/fr/oukilson/backend/dto/CreationResponseDTO.java
@@ -1,0 +1,12 @@
+package fr.oukilson.backend.dto;
+
+import lombok.*;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class CreationResponseDTO {
+
+    private boolean success;
+    private String message;
+}
