+++ src/main/java/fr/oukilson/backend/dto/UserCreationDTO.java
@@ -1,61 +1,17 @@
 package fr.oukilson.backend.dto;
 
 
+import lombok.*;
 
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.springframework.beans.factory.annotation.Qualifier;
-
-
-import java.util.regex.Pattern;
-
-
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
 public class UserCreationDTO {
 
-    private Long id;
     private String password;
     private String email;
     private String nickname;
 
-    public UserCreationDTO() {
-    }
-
-    public UserCreationDTO(Long id, String password, String email, String nickname) {
-        this.id = id;
-        this.password = password;
-        this.email = email;
-        this.nickname = nickname;
-    }
-
-
-    public Long getId() {
-        return id;
-    }
-
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getNickname() {
-        return nickname;
-    }
-
-    public void setNickname(String nickname) {
-        this.nickname = nickname;
-    }
 
 
 }
