+++ src/main/resources/application.yml
@@ -4,3 +4,8 @@ spring:
     username: root
     password: root
     driver-class-name: com.mysql.cj.jdbc.Driver
+
+environment:
+  emailRegex: "^[a-zA-Z0-9_!#$%&’*+/=?`{|}~^-]+(?:\\.[a-zA-Z0-9_!#$%&’*+/=?`{|}~^-]+)*@[a-zA-Z0-9-]+(?:\\.[a-zA-Z0-9-]+)*$"
+  nicknameRegex: "^[a-zA-Z0-9_-]{4,16}$"
+  nameRegex: "^[a-zA-Z]{2,45}$"
